:recycle: Refactor: 로그인, 닉네임 관련 재사용 가능하게 분리 완료
